CameraFollow added, new TilemapColliderTechniques and tile palette
Went a little off the theme here, but the scripts included are very reusable for the next project. The camera follow was a script i found. this map creation technique is great and might rule out needing to randomly generate a map with how quick it is using Tile Palette.